add initial version using threadLocal global bool
This version emulates try_* being expressions, but they are not. They
create an empty/default-constructed value and assigns it to the
left-hand side l-value if the input is empty, and then adds another
statment below the assignment checking the (threadlocal) global state if
it should return (empty) or not.

This has a host of issues, amongst them is that composition does not
work, and an temporary must be used to hold the evaluated result of the
input expression while deciding if it is empty or not.