refactor try_* to use GCC/Clang statement expressions
To improve overall code-quality, use the statement expression extension
in Clang/GCC to implement the try function macros. They should now be
more robust, and far more readable, and also there is no more global
variables/functions, and creating the default-constructed return value
when optional is empty is not done.